Pull from master

+++ public/pages/News/index.php
@@ -1,14 +1,14 @@
 <?php require_once(realpath(dirname(__FILE__) . '/../../..'). '/private/initialise.php'); ?>
 <?php include(SHARED_PATH . '/header.php'); ?>
 
-<link rel="stylesheet" href="/lab/stylesheets/newsStyle.css">
+<link rel="stylesheet" href="../stylesheets/newsStyle.css">
 
 <div class="header">
   <h2>News</h2>
 </div>
 
 <div class="row">
-  <div id = "main" class="leftcolumn">
+  <div id = "main" class="left">
   <?php
       function get_words($sentence, $count = 35) {
           preg_match("/(?:\w+(?:\W+|$)){0,$count}/", $sentence, $matches);
@@ -29,14 +29,16 @@
                 }
                 else{
                     echo '<div class="card">';
-                    echo '<div class="dropdown">';
-                    echo '<button class="btn btn-primary dropdown-toggle" type="button" data-toggle="dropdown"></i>';
-                    echo '<span class="caret"></span></button>';
-                    echo '<div class="dropdown-content ">';
-                    echo  '<a href="edit.php?id='.$row['articleID'].'">Edit</a>';
-                    echo '<a href="delete.php?id='.$row['articleID'].'">Delete</a>';
-                    echo '</div>';   
-                    echo '</div>';
+                    if(is_logged_in() && user_is_admin()){
+                      echo '<div class="dropdown">';
+                      echo '<button class="btn btn-primary dropdown-toggle" type="button" data-toggle="dropdown"></i>';
+                      echo '<span class="caret"></span></button>';
+                      echo '<div class="dropdown-content">';
+                      echo  '<a href="edit.php?id='.$row['articleID'].'">Edit</a>';
+                      echo '<a href="delete.php?id='.$row['articleID'].'">Delete</a>';
+                      echo '</div>';   
+                      echo '</div>';
+                    }
                     echo '<h1><a href="show.php?id='.$row['articleID'].'">'.$row['articleTitle'].'</a></h1>';
                     echo '<p>Posted on '.date('jS M Y H:i:s', strtotime($row['articleDate'])).'</p>';
                     echo '<img class="fakeimg" src="' .$row['articleImage'] .'"">';
@@ -57,17 +59,15 @@
     }
     ?>
   </div>
-  <div class="rightcolumn">
-  
-    <div class="card">
-      <h3>Popular Post</h3>
-    </div>
-    
-    <div class="card">
-     <form action="new.php">
-        <input type="submit" value="Create new article" />
-    </form>
-    
+  <div class="right">
+  <?php 
+    if(is_logged_in() && user_is_admin()){
+      echo "<div class='card'>";
+      echo "<form action='new.php'>";
+      echo "<input type='submit' value='Create new article' />";
+      echo "</form>";
+    }
+  ?>
 
     </div>
   </div>
